+++ FBSDKCoreKit/FBSDKCoreKitTests/Internal/Network/GraphRequestFactoryTests.swift
@@ -18,7 +18,7 @@ final class GraphRequestFactoryTests: XCTestCase {
       parameters: ["some": "thing"],
       tokenString: "foo",
       httpMethod: .get,
-      flags: [.skipClientToken, .disableErrorRecovery]
+      flags: [.disableErrorRecovery]
     )
     guard let graphRequest = request as? GraphRequest else {
       return XCTFail("Should create a request of the correct concrete type")